Added NS_REQUIRES_NIL_TERMINATION, NS_FORMAT_FUNCTION and NS_FORMAT_ARGUMENT macros where is is needed. Fixed a lot of format warnings/bugs due to this.

+++ Source/Additions/GCArray.m
@@ -208,8 +208,9 @@ static Class	gcClass = 0;
   if (index >= _count)
     {
       [NSException raise: NSRangeException
-		  format: @"[%@-%@]: index: %u",
-	NSStringFromClass([self class]), NSStringFromSelector(_cmd), index];
+		  format: @"[%@-%@]: index: %lu",
+	NSStringFromClass([self class]), NSStringFromSelector(_cmd),
+    (unsigned long)index];
     }
   return _contents[index];
 }
@@ -333,8 +334,9 @@ static Class	gcClass = 0;
   if (index > _count)
     {
       [NSException raise: NSRangeException
-		  format: @"[%@-%@]: bad index %u",
-	NSStringFromClass([self class]), NSStringFromSelector(_cmd), index];
+		  format: @"[%@-%@]: bad index %lu",
+	NSStringFromClass([self class]), NSStringFromSelector(_cmd),
+    (unsigned long)index];
     }
 
   if (_count == _maxCount)
@@ -419,8 +421,9 @@ static Class	gcClass = 0;
   if (index >= _count)
     {
       [NSException raise: NSRangeException
-		  format: @"[%@-%@]: bad index %u",
-	NSStringFromClass([self class]), NSStringFromSelector(_cmd), index];
+		  format: @"[%@-%@]: bad index %lu",
+	NSStringFromClass([self class]), NSStringFromSelector(_cmd),
+    (unsigned long)index];
     }
   ASSIGN(_contents[index], anObject);
   _isGCObject[index] = [anObject isKindOfClass: gcClass];
